Quick fix to remove lag..
Executing a function to make a get request while simultaneously triggering a scroll event within the executed scroll event is probably not the best idea.

This still needs work, there should probably be something else used to check for updates, such as an event where page has focus or something else that doesn't get called so many times, preferably something that reduces the timer instead.

This is just a quick fix to stop all the lag on threads when scrolling near the bottom.