Add a small margin to the off-screen post pop-up.
Without this margin you have to keep your mouse very still while
hovering a post quote for a post that is off-screen, otherwise, by
placing the mouse on the pop-up itself, the mouseover event is triggered
and the pop-up disappears, making for a very annoying user experience
when reading post pop-ups.

This patch separates the pop-up from the mouse so that this doesn't happen.