Implement priority banners.
If the directory "banners_priority" exists and contains files, a file
from it is served 33% of the time, an a file from "banners" 66% of the
time.

This may seem like an odd definition of "priority", but assuming far
fewer priority banners than normal banners, this gives them extra weight
in the selection. Before, any 1 banner in 100 had a 1% chance of being
served; now if that banner is a priority banner it has a 33% chance.